Cleaning up code
